+++ packages/convex/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @posthog/convex
 
+## 2.0.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.10
+  - posthog-node@5.36.4
+
+## 2.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.9
+  - posthog-node@5.36.3
+
 ## 2.0.13
 
 ### Patch Changes