+++ packages/next/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @posthog/next
 
+## 0.4.82
+
+### Patch Changes
+
+- Updated dependencies [[`9877710`](https://github.com/PostHog/posthog-js/commit/98777104a586651f27c0838c7377209e556d5511)]:
+    - posthog-js@1.382.0
+    - @posthog/core@1.30.10
+    - posthog-node@5.36.4
+
+## 0.4.81
+
+### Patch Changes
+
+- Updated dependencies [[`a7bd828`](https://github.com/PostHog/posthog-js/commit/a7bd828050d070e1b88eb69c3f9db71c5d08f446)]:
+    - posthog-js@1.381.0
+    - @posthog/core@1.30.9
+    - posthog-node@5.36.3
+
 ## 0.4.80
 
 ### Patch Changes