+++ packages/node/CHANGELOG.md
@@ -1,5 +1,19 @@
 # posthog-node
 
+## 5.36.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.10
+
+## 5.36.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.9
+
 ## 5.36.2
 
 ### Patch Changes