+++ packages/web/CHANGELOG.md
@@ -1,5 +1,19 @@
 # posthog-js-lite
 
+## 4.6.49
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.10
+
+## 4.6.48
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.9
+
 ## 4.6.47
 
 ### Patch Changes